Dependence Graph creates "unknown pointer node" for use of unknown pointers.
This required some refactoring of the way nodes are stored during both
dependence graph construction and in the dependence graph.

For a simplified view, we introduced the objects NodeMap,
ImmutableNodeMap and FlowDependence to a) replace long tables and
multimaps and b) also allow storage of dependence graph nodes that
are not directly related to a specific CFA edge.
This is necessary so that we can store the Unknown Pointer Node
and additional special nodes that may arise in the future.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27917 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c