Slicing Refiner determines refinement root once per target path
Previously, one refinement root was determined per slice.
This lead to a potential error when using incremental slices,
because an incremental slice for a new slicing criterion
on the same target path may be empty, leading to an error.
(It may be empty because a previous slice may already contain all
of its nodes).

In addition, it was simply more work to call it for each partial slice,
instead of the full slice for a target path

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27919 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c