Slicing Refiner supports restart strategies root and pivot.
Slicing Refiner itself does not update the precision globally anymore,
but this must be done by the calling instance.

A global update to the precision could change the precision in an
unrelated subtree and produces new inconsistent states
when a new program slice is used for a not removed state
that assumed a different slice.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27920 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c