Slicing Refiner creates incremental slice for target path subgraphs.
To compute incremental slices correctly, do not use the first
state of each target path as old precision,
but the union of the precisions of all states in the subgraph
that starts at the first state of each target path.

This new behavior does not allow us to choose whether we
want to build incremental slices. Slices always have to be computed
independent of possible refinement roots, because we don't know the new,
merged precision of the subgraph beforehand.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27921 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c