Copy ARG graphs in report into a dummy witness tab
Next step is to adapt it that it actually displays the witness, not
the ARG. This commit just ensures that we get all the mechanics right.
A lot of code was copied from the ARG tab, refactoring should be done
once it is clear which parts keep the same and which will differ.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27932 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c