Make report output actual witness data
This needed some refactoring, mainly adding the specification to the
CPAcheckerResult in order to be able to instantiate an WitnessExporter
inside the ReportGenerator. It is unfortunate that the regular witness
output happens via the CPAcheckerResult.printStatistics method while
the ReportGenerator is somehow separate. A possible future refactoring
would be to make the ReportGenerator part of the (ARG)Statistics.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27934 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c