Make witness html report feature configurable
Since the feature is not completely mature, hide it by default
behind a configuration option (report.witness, defaults to false).
Currently it is only designed for proof witnesses. It also does not
show the right witness invariants for k-induction/BMC because of the
different InvariantGenerator that is used for that analysis.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27940 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c