Add a class that executes an external ltl-to-buechi tool
The class is currently error-prone insofar, as it doesn't check whether the
external tool is actually available yet. This will be fixed soon (after
having consulted PW).

The intention of this class is to call the external tool, parse the result
(a buchi-automaton in 'Hanoi-Omega-Automaton'-format), and finally process
the returning value by building an automaton-object from the CPAchecker-
framework.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/ltl-fairness-mt@27965 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c