Reduce log level if automaton sees a failing ASSERT clause.
This message could happen a lot and confuse users, because it is printed
before the other CPAs can decide whether the current state is actually
unreachable.
Furthermore, it is inconsistent because it is triggered only via ASSERTs
in the automaton, not via other methods of going to the error state.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28172 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c