Revert a bug from r28416
That commit intended to let ExceptionHandlingAlgorithm handle
UnrecognizedCodeException (a super class of UnsupportedCodeException),
but it actually changed it to handle UnrecognizedCCodeException only,
which is not even a super class of the original class
UnsupportedCodeException.

This commit reverts the exception handling to UnsupportedCodeException
because it is not obvious whether handling UnrecognizedCodeException is
a good idea.
After all, this exception handling is enabled by the configuration
option "analysis.continueAfterUnsupportedCode",
which mentions only unsupported code.
Furthermore, no reason or other explanation (e.g., a use case) why handling
UnrecognizedCodeException in ExceptionHandlingAlgorithm should be done
if analysis.continueAfterUnsupportedCode is enabled
was given in the commit message of r28416,
and neither was the comment that explains the handling of unsupported
code updated.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28419 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c