SMG: refactoring, use SMGValues instead of Integers in all main data-structures of SMG-analysis.
Due to the nature fo this change some parts might break.
JUnit and integration tests are working.

Now we have the correct layout of SMGs,
i.e., a bipartite directed graph with real "SMGObjects" and "SMGValues".

Before all SMGValues (all types of SMGValues!) were converted to Integer
and read from Integer when needed,
into a type of SMGValue that depended on the given use-case and context.
This seemed to be very error-prone.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28452 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c