Replace CPAchecker's approximation of non-linear arithmetic with JavaSMT's
Previously non-linear arithmetic was replaced with UFs
if the solver does not support it.
The new handling in JavaSMT is more flexible and allows us to always
replace non-linear arithmetic with UFs, even if the solver would support it.
This is our new default, because when a MathSAT version with more
non-linear operations was added in r28303, the results showed that this
does not seem helpful.
(The intuition is that either linear arithmetic is sufficient,
or we should use bitvectors anyway.)

Note that this affects only configurations that use integers and
rationals, not those with bitvectors and floats.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28493 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c