Replace "assert false" with "throw new AssertionError"
There is never a reason to use "assert false".
The only reason to use asserts at all is for expensive checks
that should be disabled during production runs,
and "assert false" obviously is not an expensive check.
Furthermore, using "throw" instead allows to simplify the code
because the compiler now knows that the following code is unreachable.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28581 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c