Make predicate analysis support aliasing between signed and unsigned types.
For example, the C standard allows to cast "int*" to "unsigned int*",
but due to our type-segmented heap model was not supported so far by the
predicate analysis (accesses through an "unsigned int*" pointer would
never see what was written through an "int*" pointer and vice versa).

This fixes #294.

Note that other forms of pointer aliasing with different types
are still not supported.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28821 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c