bugfix for invalid exported counterexamples since r28830 .
PathChecker previously contained an optimization
that used the existing solver-model (from the last refinement)
and a newly created path-formula to export a counterexample with assignments.
This optimization causes the following problem:
Whenever there is a change in the encoding of the analysis
(e.g. a remapping of SSA-indices as used for BAM-recursive),
the PathChecker has to follow the same strategy to update the SSA-indizes.
This is not possible in some cases (and also error-prone).

As we already solve a separate SMT query to get a solver-model
in all cases of branching counterexample paths,
the most simplest solution is to remove the optimization completely.

Now we always solve a newly created path-formula and use its solver-model.
This change only affects single-path counterexample-traces.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28836 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c