Make SSL connections on Mac work
The previously hard-coded certificate-validation path /etc/ssl/certs
seems to not exist on Mac.
We now use the path provided by ssl.get_default_verify_paths(),
which points to the CA certificates that are installed on the system.
If this is not available, we use requests' fallback
to the certificates bundled by the certifi module.

Hopefully this now works on all platforms.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28903 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c