Ignore binary pointer operations for overflows
After careful reading of the C11(C17) standard, I conclude that
binary operations involving two pointers like p2-p1 can exhibit
undefined behavior, but are no classical arithmetic overflows.
If this changes in the future, or if the user is interested in
analyzing those expressions anyway, it can be enabled with the
new configuration option "overflow.trackPointers".

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28974 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c