SymEx: Express def assignments in constraints formula through conjunction again
Previously, we used a formula of the structure
formula \land D \implies definites
and provided assumption D == true for the SAT check.
But it is a) errorprone to miss providing that assumption,
and b) only expresses the intended meaning (that definites always true)
if the assumption is provided.

Instead, go back to the old way of simple concatenation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29248 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c