Change return type of several methods in MachineModel-class to 'BigInteger'
The changes were carried out in a very conservative way, i.e. only
as few things as possible have been changed in order to fix the
problem as described in issue #426 on gitlab.

Each method outside of 'MachineModel', which requires the returning
type to be in int or long, now converts the BigInteger back afterwards,
with either of the BigInteger-methods 'intValueExact()' or
longValueExact()'. An exception is thereby thrown in case an overflow
occurs.

As the returning types have been int's and long's until now, I'm
expecting this convertion to not lead to any bigger problems for now.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/size-type-refactoring@29250 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c