ConstraintsCPA checks trivial constraints individually
Never adds them to the ConstraintsState.

A similar check already existed and was removed in 186cdd03aa as
a simplification,
but this previous check was more elaborate and took the definite
assignments of constraints into account.

The current check only checks for constraints without any
symbolic identifiers (e.g., 5 > 4), checks these immediately
and doesn't add them to the constraints state.

The reason behind this is that such constraints cluttered the
constraints state and worsened readability.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29552 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c