Consider definitions of assignments in dependece graph to get a consistent graph
Otherwise, nodes with and without the dependence may exist that actually have the same
meaning but different edges between each other

Example:

A node for a CFA edge g that is flow dependent on another CFA edge h will
look for a dependence-graph node that represents h with the variable that is used in g and defined in h.
But if we don't consider definitions during flow-dependece collection,
it will not find one and create a new node that misses all further dependences.
We also can't discard the definitions all-together because this would make
interprocedural slicing less concrete (whenever an edge is dependent on a function parameter,
it would be dependent on all function parameters according to the dependence graph)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29553 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c