Fix int literals to be parsed with their correct types
According to the C standard sect. 6.4.4.1, the type of an integer constant is
the first on a list in which its value can be represented. The according list
can be found in the beforementioned section. This has not been correctly
implemented in CPAchecker previously, thus leading to issues such as #435
and #455.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29588 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c