Do not create invariants over unsupported types (such as arrays)
Such invariants do not make sense, and lead to crashes when converted to
SMT formulas.
In InvariantsTransferRelation.handleDeclaration there was already this
check for supported types, I copied it to handleFunctionCall,
which also handles declarations (of formal parameters).
Hopefully there is no other code that also handles declarations.

This should fix #477.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29672 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c