Fix a problem in ParallelAlgorithm that led to unchecked exceptions being swallowed.
The class CPAchecker contained a workaround for this problem,
but this had no effect in more complex scenarios,
such as a ParallelAlgorithm wrapped in another algorithm.

So now we directly propagate all unchecked and unhandled exceptions.
This will probably lead to cases where expected but unhandled exceptions
trigger crashes, but then we should fix these cases specifically.

The exception handling of the ParallelAlgorithm needs general
improvement anyway (cf. #326).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29708 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c