Change parsing of float-literals and add some tests
The changes fix the issue described in #492, however, this subsequently leads to another
problem which is described in #500 (that is, there is currently no support for 'long
double').

I have also moved the 'adjustPrecision'-method as a temporary fix to the ASTLiteralConverter.
This code should be completly unnecessary once the BigDecimals are properly replaced by
CFloats in 'AFloatLiteralExpression'-class.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29716 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c