Improve value visitor's cast functionality and support long double
Generalize and simplify code for value casts in value analysis,
especially handling of different integer types;
also handle cast to long double, if possible.
If the value that is supposed to be cast to a long double fits
precisely into a double, handle the cast.
Otherwise, return value 'unknown', as is done in many other
methods of the value visitor.

Solves issue #500

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29749 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c