PredicateCPARefiner: bugfix for optimization of branchingformulas.
Some paths contain branching in a way that checking
the number of states along the path is not sufficient.
We need to check each single state whether is leads to
branching along the path itself, as it can happen for
ConditionTrees in the CFA.

Example:
A path contains states A, B, C with successor relations
A->B, B->C, A->C. This branching was not be detected before.

See #517 for more detail.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29779 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c