Add option for treating imprecise counterexamples of PredicateCPA as a bug.
Most of the time, these are bugs, and continuing hides these bugs.
In some cases, however, such imprecise counterexamples are expected,
for example, for the UF-based heap encoding (cf. #283).
Thus we need an option that keeps the current behavior.

For now, the default is the current behavior,
hopefully we can switch this soon.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29840 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c