Update some dependencies, including Guava and Truth
To do this we need to disable null checks of SpotBugs,
but there is no progress in SpotBugs on supporting the new annotations,
and we do not want to use the old dependencies infinitely.

Upgrading Truth means that we need to refactor the subject implementation
that we have for AST matchers because of an API deprecation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29918 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c