Migrate most annotations like @Nullable away from javax.annotation (jsr305.jar)
This packages creates problems with the module system of Java 9,
and Guava and other libraries have also started migrating away from it.
We now use it only for `@ParametersAreNonnullByDefault`, like Guava.
We also keep a few cases of `javax.annotation.concurrent.Immutable`.
This annotation basically has no effect and serves only as
documentation.
We cannot migrate to the error-prone version of `@Immutable`
because then error-prone would attempt to prove these classes as actually
immutable, and fail.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@29943 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c