Revert r24366 (36bf935e), r24373 (e11c7384), and r24374 (b5733204): fix witness and counterexample export that was broken by r29988 (53d381fb)
r29988 reverted the heap modelling of the predicate analysis from a
bit-adressable heap to a standard byte-addressable heap.
It turns out that all the common code of the counterexample and witness
export encoded the assumption previous of the predicate analysis
that all addresses are in bits, although this was not documented.
Thus, after removing this assumption from the predicte analysis,
the address computations used for exporting heap values for
counterexamples of the predicate analysis was broken.

Furthermore, because the common code assumed a bit-addressable heap,
even the value analysis had been changed to also use this assumption of
the predicate analysis.

This commit now reverts the commits that added the bit-addressable heap
to the common counterexample-export code and the value analysis.
Hopefully now everything is restored to assume a standard
byte-adressable heap.
However, as there are no tests for this, it is hard to say whether
this is complete, or whether other CPAs such as the InvariantsCPA or
SMGCPA where also changed in the meantime to assume a bit-addressable heap.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30007 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c