Remove option cpa.callstack.depth from configs for k-induction and PDR
Setting this option is unsound for most configurations of CPAchecker,
except if specific support for recursion is present
(usually using BAM).
This is not the case for k-induction and PDR.

It is unclear why this option was even there in the first place,
it was added when the single-loop transformation was disabled,
but this shouldn't cause the existence of recursion.
(In the PDR config it is obviously just a result of copy-and-paste.)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30021 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c