Fix problem with erroneous recursion detection in step case of induction
Because we have initial states in the middle of functions,
we can reenter these functions on the same path later on
(via a loop in a calling function, for example),
and this is erroneously detected as recursion.
This happens only in the step case and until r30021 (dd809f04)
was circumvented by setting cpa.callstack.depth=1,
but this is unsound.
So now we fix the logic in the callstackCPA instead.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30024 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c