Add values to AlgorithmStatus and Result enum for cases without a checked property
We can use this now to use a NoopAlgorithm that does nothing
instead of adding a special case for the empty analysis to the
CPAchecker class.
The end result of CPAchecker will then not be TRUE/FALSE/UNKNOWN,
but simply "Finished".

Other algorithms should also return this value in the future
if TRUE/FALSE/UNKNOWN doesn't make sense for them.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30038 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c