Fix unintuitive option handling in configViz.py
If --dir DIRECTORY [DIRECTORY ...] is specified on the command line,
the user does not expect this to be just appended to the default
values. There would also be no way to exclude test/config from the
generated graph. It is better to replace the default values with the
list explicitly supplied by the user.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30061 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c