Refactor in CPA-w2t: Function/variable names and docstrings
Rename function and variable names related to compilation
to reflect correct meaning. before, only gcc was used,
which was reflected in the names - but now,
we can also use clang, so make the function/variable names
more generic.

Add docstrings to document behavior and return values of some
not-so-obvious functions.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30109 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c