Change successor to be more merge-friendly
The getAbstractSuccessor-method is only called if there is no
LocationCPA. UnknownAutomatonState has the downside that it remembers
all of its previous states, which leads to never reaching a fixed point
because successors can never be covered/merged by their predecessors.
The overapproximation with the top state is a much better choice here.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/slab@30161 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c