Make (overflow) precondition handling more generic
This is a refactoring. It removes dependencies between the packages
cpa.overflow and cpa.predicate. A AbstractStateWithAssumptions will
now simply provide a set of the (previous) abstract states for which
the precondition assumptions should hold.

It is now the callers responsibility to extract the right information
from that set. The predicate analysis e.g. will extract the
PredicateState and use its PathFormula for applying the preconditions.

This will allow other analysis like the value analysis to support
overflows in a similar fashion.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30376 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c