Enhance producer information in witness files
If the producer field only lists CPAchecker and its version, it is
hard to distinguish witnesses produced by different tools or by
witness validation from another. Now we will add the name of the
configuration used to the string by default. The user can also
configure this substring by using the option analysis.name
either in their configuration or on the command line with -setprop.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30456 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c