Refactor new producer information for general use
The additional information about the approach that is executed is
useful in different contexts, therefore it is refactored out of the
witness context. As discussed with @dbeyer, this information is now
also displayed in the standard output along the CPAchecker version,
c.f. issue #550 for further background about this topic.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30458 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c