Generalize stengthening of predicate + assumptions
There was already a fix present for handling AssumptionStorageStates.
But the PredicateTransferRelation has also other cases where it
extracts assumptions (or formulas) from other states. For example,
retrieving assumptions from an AutomatonState did NOT work correctly
if the transition also contains a check (e.g. CPAQuery for location),
since the assumptions are hidden in an AutomatonUnknownState until
it is strenghtened into a normal AutomatonState.
I already tested this with integration-predicate-bitprecise.xml and
will have a look whether this breaks something for the other tests.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30583 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c