Automaton for validating correctness witness
An automaton is built as additional verfication goal based
on the invariants in the given correctness witness. The
automaton is as ControlAutomatonCPA part of the CPAs
in the CEGAR algorithm. The specification of the
automaton is then checked using predicate abstraction.
This is the first approach so later changes are likely.
Commit includes also some name refactorings for clarity.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/correctnessWitnessValidationUsingPredicateAnalysis@30639 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c