SSC: change path computation towards a lazier strategy.
Before, we computed full-paths directly and used them for export.
Now we compute them only on demand, which safes time on larger examples
with counterexamples that are filtered before export.

This commit breaks the SSC-refinement-related code,
i.e., updating or removing ARG-subtrees is no longer possible.
If needed, it has to be re-added again.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30642 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c