Add handling of many children to ARGTOCTranslator
Before, we were not able to generate programs from an ARG that had
states with more than two children. This will now be converted into
a nondeterministic branching, where the conditions are handled with
__VERIFIER_assume(cond). We now also handle assumptions from states
implementing AbstractStateWithAssumption in an analogous manner.
This commit comes with basic tests for the new features.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30653 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c