Fix bad algorithmic performance of a method
There is no need to backtrack the list of seen elements. Once a element
is seen and all its children were followed, it is clear that the
initial state can either be reached from that state or not. So if we
reach it via another branch after backtracking, we do not need to
consider it, because having backtracked before means that the state
is in the set of states from which the initial state is not reachable.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30720 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c