Avoid accidental correctness witness generation
Before this fix it could happen (e.g. for k-Induction) that a
subanalysis generates a correctness witness despite the overall result
being FALSE. This is because a correctness witness is also generated
if the result is UNKNOWN, and results for all subanalyses of an
ParallelAlgorithm that are not the successful analysis are set to
UNKNOWN. This commit changes the latter to DONE, which will then in
return not trigger the correctness witness generation.
The result value DONE was probably not available when the code for the
subanalyses' result was written, it seems like a much better choice
here than using UNKNOWN.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/witness2test_jakob@30780 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c