CFAToCTranslator closes branches at join nodes
If branches join, it is checked whether they have a common ancestor
outer scope, and if they do, the outer scope if used for following
commands. To make this work, traverse through the CFA in a (simulated)
reverse post-order by only considering CFA nodes and edges
for which all predecessors have already been handled.
We don't handle loops this way, but still use gotos there,
because it would result in a dead-lock, otherwise.

This commit also adjusts the exact location of block starts:
Previously, blocks of branches started at the branching node.
Now, blocks of branches start with the first node after the branching
condition.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/portfolio_testcase_generation@30801 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c