Let CFAToCTranslator work on CFA nodes instead of edges
This simplifies multiple parts of the translator
and fixes a bug that I didn't know how to fix before:
BFS traversal through the CFA is now possible with correct re-use
of already visited edges. Before, if three or more edges
were used in a join, all but the last would miss the go-to statement
because the node was not considered, yet.
By working on nodes, only, all of this is simplified.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/portfolio_testcase_generation@30808 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c