Improve parsing of larger witnesses significantly
The call to getLength apparently is not cached. This meant that
if we iterate over the states of a large witness with asIterable,
for each iteration the whole DOM is (very slowly) traversed.
This scales so badly (quadratic on 100_000 nodes, do the math)
that parsing such a witness was practically impossible.
Now we simply cache the length and curse the org.w3c.dom API.
Length calculation is done lazily to avoid unnecessary work.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30841 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c