CFA with Enums: allow different types for different enumerations.
Previously we only had SIGNED_INT for enums.
As C99 allows all integer types for enums,
we have to also support longer types.
This commit sets the correct type for enums in the CFA
and is a large step towards a solution for #348.

In some future commits we will update analyses
that handle enums and use their type, such as predicate analysis.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30847 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c