Use outer-most branch for code after join nodes in CfaToCTranslator
The previous technique on join-nodes put code after join nodes
in the first block that reached the join node based on the traversal
strategy.
This could lead to very deep nesting levels (> 2500 on ECA).
To counter this, we first tried to compute common surrounding blocks,
but this didn't work for complicated cases (e.g., as soon as a goto
exists, or conditional or's)

Now, we first try to merge if-else constructs based on their surrounding
block, and then put code into the code block with the smallest nesting
level.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/portfolio_testcase_generation@30901 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c