Fix broken corr. witness generation of k-induction
I broke this with r30717. With k-induction, the correctness witness
is generated by the invariant-generation analysis, which actually
has no way of knowing whether k-induction proved the program to be
correct or not, so it did just output a correctness witness EVERY
time. For verdict FALSE, this was then overwritten by the violation
witness (if both file names match). With r30717, I tried to avoid the
useless, expensive generation of a correctness witness in the FALSE
case. I did not realize that this meant that k-induction whould not
output a correctness witness in the TRUE case. Now this commit should
fix that. This still has a certain code smell, but a better solution
would probably require larger changes to the code base.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30911 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c