Parser: always simplify constant enum expressions.
We need the integer value for a correct enum type.
Until now, the enum value was ignored (and set to NULL),
when options like 'cfa.simplifyConstExpressions' were disabled.

This was sound, but possibly not the best option.

Now, we always simplify constant enum expressions.
This might lead to problems (unsoundness) with specifications
that check for special operations, like overflows, because these
configurations have no possibility to look at the computation
(and simplification) of an enum value.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@30913 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c