CFloatImpl correctly computes decimal representation for floats < 1
During computation of decimal positions,
underflows from exponent to mantissa were not considered.

Example:

Float-value 0.6
=> Exponent: -1, Mantissa = 1.2

Computation of decimal positions considers integral and fractional
separately, and first computes 1 * 2^-1 = 0.
Afterwards, 0.2 * 2^-1 = 0.1 was computed,
and the (wrong) result was 0.1 .

Correct this by checking whether the integral is 1
and adding the remainder (1 * 2^-1 = 0.5) to the fractionals.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31017 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c