ValueAnalysis does not track any addressed variable or their members
Previously, members of addressed variables were still tracked,
and only addressed pointers were put on the blacklist.
Now, we don't track any member if the variable itself is on the
blacklist, and we put all addressed variables on the blacklist.

Note that this may be too strict and we should check that strengthening
from pointer analysis is still used properly.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31020 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c