Let CFA->C translator use target labels
If the CFA goes to a node that is directly followed
by a blank edge "LABEL: x" that goes to a CLabelNode,
use a goto with label x, instead of looking for
previous labels.

This fixes an issue with PROG_EXIT labels
preceded with BLOCK-labels.
In such a case, goto-statements that targeted the
PROG_EXIT label used the BLOCK labels in the resulting
C code, which led to correct, but confusing program code.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/portfolio_testcase_generation@31096 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c