ASTLiteralConverter: Add an additional filter to possible suffix candidates
This fixes issue #570.

According to the C standard, if no u/U-suffix is given in the source-code for
an integer constant, the actual type may then under no circumstances be
represented by an unsigned int type either. However, this affects only decimal
constants, i.e. octal and hexadecimal constants are unaffected by this rule.

The same is furthermore true for the opposite, i.e. integer types that are
declared as unsigned may neither be represented by signed integer types.

This filter was missing previously (due to an oversight), and should fix the
issue as described in #570.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31117 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c