ASTLiteralConverter: Fix erroneous unit test
Integers not explicitely declared as unsigned must not be
extended to unsigned integer constants while trying to match
them to their smallest matching type (this holds only
for decimal integer constants).
As a result, an int value greater than 2^63-1 may not be matched
to an unsigned int when given as a decimal number.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31136 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c