fix exception handling in ParallelAlgorithm
createParallelAnalysis produced Callables that throw
InvalidConfigurationException, but handleFutureResults does not handle them.
This lead to stack traces shown for invalid configurations,
for example if -default was used on Mac or Windows.

Now ParallelAlgorithm creates all the nested analyses in its constructor
and InvalidConfigurationExceptions are propagated normally.
As part of #318, we also abort immediately if one of the analyses
throws a CPAException from its constructor instead of continuing
to run the other analyses, often swallowing the exception completely.

This means that less work is done in parallel,
but constructing analyses should be cheap compared to running them anyway.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31192 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c