BAM: possilty fix an exception when removing subtrees from ARGs.
When removing an already destroyed state from an ARG,
BAM (with ARGInPlaceSubtreeRemover) fails with an exception.
This commit avoids the exception and skips the removing process in this case.

The bug/exception appears for the following cmdline:
scripts/cpa.sh -default--overflow -64 \
    -spec test/programs/benchmarks/properties/no-overflow.prp \
    test/programs/benchmarks/termination-crafted/NestedRecursion_1d.c

The cmdline from above runs into timeout (at least 100s) now.
It seems to have progress (increasing BAMrec fixpoint iterations),
so probably it is no endless loop.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@31216 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c